+++ services/iotcentral/mgmt/2018-09-01/iotcentral/CHANGELOG.md
@@ -1,4 +1,4 @@
-Generated from https://github.com/Azure/azure-rest-api-specs/tree/3c764635e7d442b3e74caf593029fcd440b3ef82/specification/iotcentral/resource-manager/readme.md tag: `package-2018-09-01`
+Generated from https://github.com/Azure/azure-rest-api-specs/tree/4c93f28f89435f6d244f4db61bbf810b5d20f09f/specification/iotcentral/resource-manager/readme.md tag: `package-2018-09-01`
 
 Code generator @microsoft.azure/autorest.go@2.1.168
 
@@ -14,6 +14,7 @@ Code generator @microsoft.azure/autorest.go@2.1.168
 
 ### Removed Struct Fields
 
+1. AppTemplate.AppTemplateName
 1. AppsCreateOrUpdateFuture.azure.Future
 1. AppsDeleteFuture.azure.Future
 1. AppsUpdateFuture.azure.Future
@@ -22,6 +23,7 @@ Code generator @microsoft.azure/autorest.go@2.1.168
 
 ### New Struct Fields
 
+1. AppTemplate.Name
 1. AppsCreateOrUpdateFuture.Result
 1. AppsCreateOrUpdateFuture.azure.FutureAPI
 1. AppsDeleteFuture.Result